rename:study-main내 todo 관련 todo 폴더 생성
